refactor: model and data
